Bug fixes
- Added foreign keys to student migration
- Rejected student proposals are now deleted
- Updated foreign keys for onDelete and onUpdate
- Updated intrested students project proposal table
- Updated mail:to calls to catch exceptions (Added user dialog where needed)
- Fixed project web routes being in incorrect middleware
- Fixed a bug where students could edit a proposed project
- Fixed a bug where students could edit project status
- Fixed a loader bug in the student project preview
- Fixed a header bug where accepted project was not shown
- Fixed a bug in main.js where an error would thrown when user wasn't on the main page

+++ app/resources/views/projects/project.blade.php
@@ -92,19 +92,15 @@
 			@endif
 		@endif
 		
-		@if($project->isOwnedByUser())
+		@if($project->isOwnedByUser() && !Auth::user()->isStudent())
 			<a class="button button--raised" href="{{ action('ProjectController@edit', $project->id) }}">Edit Project</a>
 
-			<form class="delete-project" action="{{ action('ProjectController@destroy', $project->id) }}" data-project-title="{{ $project->title }}" method="DELETE" accept-charset="utf-8">
-				<button type="submit" class="button button--raised button--danger" title="Delete {{ $project->title }}">Delete Project</button>
-			</form>
 			{{-- <form class="delete-project" action="{{ action('ProjectController@destroy', $project->id) }}" data-project-title="{{ $project->title }}" method="DELETE" accept-charset="utf-8">
-				<button type="submit" class="button button--svg button--danger" title="Delete {{ $project->title }}">
-					@include('svg.bin')
-					<p>Delete</p>
-				</button>
+				<button type="submit" class="button button--raised button--danger" title="Delete {{ $project->title }}">Delete Project</button>
 			</form> --}}
 		@endif
+
+
 
 	</div>
 </div>
